email/requirements.txt
get_docker_secret
psycopg2-binary==2.8.5
